提交了修改后的代码

+++ src/main/java/com/newframe/entity/user/CapitalFlow.java
@@ -153,14 +153,14 @@ public class CapitalFlow {
      * ctime
      */
     @Column(name = "ctime")
-    private Long ctime;
+    private Integer ctime;
 
     /**
      * utime
      * utime
      */
     @Column(name = "utime")
-    private Long utime;
+    private Integer utime;
 
     public CapitalFlow() {
     }
